backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
email-validator>=2.1.0
python-slugify>=8.0.1
aiohttp>=3.9.0
asyncpg>=0.29.0
